pyproject.toml
[project]
name = "complianceqapipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.129.0",
    "groq>=1.0.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.8",
    "openai-whisper>=20250625",
    "pandas>=3.0.0",
    "pydantic>=2.12.5",
    "pypdf>=6.7.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "uvicorn>=0.41.0",
    "yt-dlp>=2026.2.4",
]
